backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
groq==1.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic>=2.0
email-validator==2.3.0
httpx
sendgrid==6.11.0
